packages/common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "quantike", email = "55246750+quantike@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.2",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
